backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-multipart
pdfminer.six
python-docx
pandas
nltk
aiofiles
rapidfuzz
sentence-transformers
faiss-cpu
Cython